Start refactoring and code cleaning, There is several coding convention that are not respected :
1- Never use unsigned long cast for pointer if it is not absolutly necessary
2- Always use references when an object given in parameter must exist
3- Always indicates in the documentation who is the owner of the created object (needed for memory management)
4- Try to limit the link between objects when it is possible 



git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@3562 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/src/GlSelectSceneVisitor.cpp
@@ -28,7 +28,7 @@ namespace tlp {
 
   void GlSelectSceneVisitor::visit(GlSimpleEntity *entity) {
     if(selectionFlag==SelectSimpleEntities)
-      calculator->addSimpleEntityBoundingBox((unsigned long)entity,entity->getBoundingBox());
+      calculator->addSimpleEntityBoundingBox(entity,entity->getBoundingBox());
   }
 
   void GlSelectSceneVisitor::visit(GlNode *glNode) {
